architectural change for APIV3
