+++ library/src/main/java/net/pubnative/library/request/model/PubnativeAdDataModel.java
@@ -1,0 +1,8 @@
+package net.pubnative.library.request.model;
+
+import java.util.HashMap;
+
+public abstract class PubnativeAdDataModel {
+    protected String                    type;
+    protected HashMap<String, String>   data;
+}
