+++ library/src/main/java/net/pubnative/library/request/model/PubnativeExtModel.java
@@ -1,0 +1,8 @@
+package net.pubnative.library.request.model;
+
+import java.util.List;
+
+public class PubnativeExtModel {
+
+    protected List<PubnativeAPIV3ExtModel> extData;
+}
